@@ -1,16 +1,13 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
 """The setup script."""
 
 __author__ = "Ag Stephens"
 __contact__ = "ag.stephens@stfc.ac.uk"
 __copyright__ = "Copyright 2018 United Kingdom Research and Innovation"
-__license__ = "BSD - see LICENSE file in top-level package directory"
 __version__ = "0.1.0"
 
-from setuptools import find_packages
-from setuptools import setup
 import os
+
+from setuptools import find_packages, setup
 
 # One strategy for storing the overall version is to put it in the top-level
 # package's __init__ but Nb. __init__.py files are not needed to declare
@@ -38,42 +35,29 @@ setup(
     # https://www.python.org/dev/peps/pep-0301/#distutils-trove-classification
     classifiers=[
         "Development Status :: 5 - Production/Stable",
-        "Environment :: Console",
-        "Environment :: Web Environment",
-        "Intended Audience :: End Users/Desktop",
         "Intended Audience :: Developers",
-        "Intended Audience :: System Administrators",
         "Intended Audience :: Science/Research",
         "License :: OSI Approved :: BSD License",
         "Natural Language :: English",
-        "Operating System :: Microsoft :: Windows",
-        "Operating System :: POSIX :: Linux",
+        "Operating System :: OS Independent",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
-        "Topic :: Security",
-        "Topic :: Internet",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
         "Topic :: Scientific/Engineering",
-        "Topic :: System :: Distributed Computing",
-        "Topic :: System :: Systems Administration :: Authentication/Directory",
         "Topic :: Software Development :: Libraries :: Python Modules",
     ],
     description="Grid definitions for the roocs regridder",
-    license=__license__,
-    # This qualifier can be used to selectively exclude Python versions -
-    # in this case early Python 2 and 3 releases
-    python_requires=">=3.6.0",
+    license_files=["LICENSE"],
+    python_requires=">=3.9.0",
     install_requires=requirements,
     long_description=_long_description,
     long_description_content_type="text/x-rst",
     include_package_data=True,
     keywords="roocs_grids, regrid, grid, interpolate, cmip",
     name="roocs_grids",
-    packages=find_packages(
-        include=["roocs_grids"]
-    ),
+    packages=find_packages(include=["roocs_grids"]),
     setup_requires=setup_requirements,
     test_suite="tests",
     tests_require=test_requirements,
